refactor: consolidate useUserContext
