--- src/hooks/useUserContext.jsx
@@ -1,7 +1,0 @@
-import { useContext } from 'react';
-import { UserContext } from '../context/userContext';
-
-
-export const useUserContext = () => {
-  return useContext(UserContext);
-};
